+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## 0.3.2 (TBA)
 
--   Chat message objects have a new `html` field ([docs](./docs/rooms.md#messaging)).
+-   Chat message objects have a new `html` field ([docs](./docs/messages.md)).
 
 ## 0.3.1 (2022-04-30)